pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "standstill"
version = "0.1.0"
description = "AWS Control Tower management CLI"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = ["aws", "control-tower", "security", "organizations", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer>=0.12,<0.16",
    "rich>=13,<15",
    "boto3>=1.34,<2",
    "botocore>=1.34,<2",
    "pyyaml>=6,<7",
    "pydantic>=2,<3",
]

[project.urls]
Homepage = "https://github.com/dbnz-io/standstill"
Repository = "https://github.com/dbnz-io/standstill"
Issues = "https://github.com/dbnz-io/standstill/issues"
Changelog = "https://github.com/dbnz-io/standstill/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "moto[controltower,organizations,iam,sts]>=5,<6",
    "pytest-cov>=5,<7",
    "ruff>=0.4,<1",
]

[project.scripts]
standstill = "standstill.main:app"

[tool.hatch.build.targets.wheel]
packages = ["standstill"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 130

[tool.ruff.lint]
select = ["E", "F", "I"]
